+++ package.json
@@ -10,17 +10,17 @@
   },
   "dependencies": {
     "@connext/sdk": "^2.1.1",
-    "@daohaus/connect": "^0.2.0-alpha.0",
-    "@daohaus/contract-utils": "^0.2.0-alpha.0",
-    "@daohaus/form-builder": "^0.2.0-alpha.0",
-    "@daohaus/keychain-utils": "^0.2.0-alpha.0",
-    "@daohaus/moloch-v3-fields": "^0.2.0-alpha.0",
-    "@daohaus/moloch-v3-hooks": "^0.2.0-alpha.0",
-    "@daohaus/moloch-v3-legos": "^0.2.0-alpha.0",
-    "@daohaus/moloch-v3-macro-ui": "^0.2.0-alpha.0",
-    "@daohaus/tx-builder": "^0.2.0-alpha.0",
-    "@daohaus/ui": "^0.2.0-alpha.0",
-    "@daohaus/utils": "^0.2.0-alpha.0",
+    "@daohaus/connect": "^0.2.1",
+    "@daohaus/contract-utils": "^0.2.1",
+    "@daohaus/form-builder": "^0.2.1",
+    "@daohaus/keychain-utils": "^0.2.1",
+    "@daohaus/moloch-v3-fields": "^0.2.1",
+    "@daohaus/moloch-v3-hooks": "^0.2.1",
+    "@daohaus/moloch-v3-legos": "^0.2.1",
+    "@daohaus/moloch-v3-macro-ui": "^0.2.1",
+    "@daohaus/tx-builder": "^0.2.1",
+    "@daohaus/ui": "^0.2.1",
+    "@daohaus/utils": "^0.2.1",
     "@tanstack/react-table": "^8.7.9",
     "cross-env": "^7.0.3",
     "moment": "^2.29.4",
