+++ src/components/registries/ForeignRegistryOverview.tsx
@@ -1,8 +1,5 @@
 import { AddressDisplay, H4, DataIndicator, Tag } from "@daohaus/ui";
-import {
-  ZERO_ADDRESS,
-  formatShortDateTimeFromSeconds,
-} from "@daohaus/utils";
+import { ZERO_ADDRESS, formatShortDateTimeFromSeconds } from "@daohaus/utils";
 import { Keychain } from "@daohaus/keychain-utils";
 import { REGISTRY, TARGETS } from "../../targetDao";
 import { RegistryMenu } from "./RegistryMenu";
@@ -12,20 +9,15 @@ import {
   RegistryCardHeader,
   RegistryOverviewCard,
 } from "./RegistryOverview.styles";
+import { RiCheckboxCircleFill } from "react-icons/ri";
 
 type RegistryProps = {
   target?: REGISTRY;
-  owner?: string;
-  lastUpdate?: number;
-  totalMembers?: number;
   foreignRegistry?: REGISTRY;
 };
 
-export const RegistryHomeOverview = ({
+export const ForeignRegistryOverview = ({
   target,
-  owner,
-  lastUpdate,
-  totalMembers,
   foreignRegistry,
 }: RegistryProps) => {
   const daochain = TARGETS.NETWORK_ID;
@@ -45,36 +37,41 @@ export const RegistryHomeOverview = ({
               }
               truncate
               copy
-              explorerNetworkId={daochain as keyof Keychain}
+              explorerNetworkId={
+                (foreignRegistry?.NETWORK_ID as keyof Keychain) ||
+                (daochain as keyof Keychain)
+              }
             />
 
-            <Tag tagColor="pink">Home</Tag>
+            <Tag tagColor="blue">Foreign</Tag>
+            {foreignRegistry?.REGISTRY_ADDRESS != ZERO_ADDRESS ? (
+              <Tag tagColor="green">
+                {/* registered? */}
+                <RiCheckboxCircleFill />
+              </Tag>
+            ) : (
+              <Tag tagColor="red">
+                <RiCheckboxCircleFill />
+              </Tag>
+            )}
           </TagSection>
         </div>
         <div className="right-section">
-          <RegistryMenu home={true} foreignRegistry={foreignRegistry} />
+          <RegistryMenu home={false} foreignRegistry={foreignRegistry} />
         </div>
       </RegistryCardHeader>
       <DataGrid>
         <>
           <DataIndicator
             label="Total Members"
-            data={totalMembers?.toString() || "NA"}
+            data={foreignRegistry?.TOTAL_MEMBERS?.toString()}
           />
+          <DataIndicator label="Status" data={"..."} />
+
           <DataIndicator
-            label="Owner"
-            data={
-              owner?.toLowerCase() == TARGETS.SAFE_ADDRESS.toLowerCase()
-                ? "DAO"
-                : owner?.toString()
-            }
-          />
-          <DataIndicator
+            label={`Last Sync`}
             size="sm"
-            label={`Last Update`}
-            data={
-              formatShortDateTimeFromSeconds(lastUpdate?.toString()) || "NA"
-            }
+            data={formatShortDateTimeFromSeconds(foreignRegistry?.LAST_ACTIVITY_UPDATE) || "NA"}
           />
         </>
       </DataGrid>
