+++ src/pages/Home.tsx
@@ -1,29 +1,18 @@
 import styled from "styled-components";
-
-import { SingleColumnLayout, Spinner } from "@daohaus/ui";
 import { useDHConnect } from "@daohaus/connect";
-import { useMemberRegistry } from "../hooks/useRegistry";
 import { useCurrentDao } from "@daohaus/moloch-v3-hooks";
+import { SingleColumnLayout, Spinner } from "@daohaus/ui";
+
 import { MemberRegistry } from "../components/MemberRegistry/MemberRegistry";
+import { useMemberRegistry } from "../hooks/useRegistry";
 import { TARGETS } from "../targetDao";
+import { HAUS_RPC } from "../utils/keychain";
 
 const LinkBox = styled.div`
   display: flex;
   width: 50%;
   justify-content: space-between;
 `;
-
-export const HAUS_RPC = {
-  "0x1": `https://787b6618b5a34070874c12d7157e6661.eth.rpc.rivet.cloud/`,
-  "0x5": `https://787b6618b5a34070874c12d7157e6661.goerli.rpc.rivet.cloud/`,
-  "0x64": "https://rpc.gnosischain.com/",
-  "0xa": "https://mainnet.optimism.io",
-  "0x89": "https://polygon-rpc.com/",
-  "0xa4b1": "https://arb1.arbitrum.io/rpc",
-  "0xa4ec": "https://forno.celo.org",
-  "0x1a4": "https://goerli.optimism.io",
-  "0x66eed": "https://rpc.goerli.arbitrum.gateway.fm"
-};
 
 export const Home = () => {
