+++ src/pages/ReplicaConfig.tsx
@@ -2,37 +2,36 @@ import React, { useEffect } from "react";
 
 import { ParLg,  SingleColumnLayout } from "@daohaus/ui";
 
+import { useParams } from "react-router-dom";
+import { create, SdkConfig } from "@connext/sdk";
 import { FormBuilder } from "@daohaus/form-builder";
+import { MolochFields } from "@daohaus/moloch-v3-fields";
+import { fromWei, ZERO_ADDRESS } from "@daohaus/utils";
+
+import { useConnext } from "../hooks/useConnext";
+import { useMemberRegistry } from "../hooks/useRegistry";
+import { AppFieldLookup } from "../legos/fieldConfig";
 import { APP_FORM } from "../legos/forms";
 import { TARGETS } from "../targetDao";
+import { HAUS_RPC } from "../utils/keychain";
 
-import { create, SdkConfig } from "@connext/sdk";
-import { fromWei, ZERO_ADDRESS } from "@daohaus/utils";
-import { useParams } from "react-router-dom";
-import { MolochFields } from "@daohaus/moloch-v3-fields";
-import { AppFieldLookup } from "../legos/fieldConfig";
-import { useConnext } from "../hooks/useConnext";
-import { HAUS_RPC } from "@daohaus/keychain-utils";
-import { useMemberRegistry } from "../hooks/useRegistry";
-
-
-const sdkConfig: SdkConfig = {
-  signerAddress: "0x2b8aA42fFb2c9c7B9f0B1e1b935F7D8331b6dC7c",
-  // Use `mainnet` when you're ready...
-  network: "testnet",
-  // Add more chains here! Use mainnet domains if `network: mainnet`.
-  // This information can be found at https://docs.connext.network/resources/supported-chains
-  chains: {
-    1735353714: {
-      // Goerli domain ID
-      providers: ["https://rpc.ankr.com/eth_goerli"],
-    },
-    1735356532: {
-      // Optimism-Goerli domain ID
-      providers: ["https://goerli.optimism.io"],
-    },
-  },
-};
+// const sdkConfig: SdkConfig = {
+//   signerAddress: "0x2b8aA42fFb2c9c7B9f0B1e1b935F7D8331b6dC7c",
+//   // Use `mainnet` when you're ready...
+//   network: "testnet",
+//   // Add more chains here! Use mainnet domains if `network: mainnet`.
+//   // This information can be found at https://docs.connext.network/resources/supported-chains
+//   chains: {
+//     1735353714: {
+//       // Goerli domain ID
+//       providers: ["https://rpc.ankr.com/eth_goerli"],
+//     },
+//     1735356532: {
+//       // Optimism-Goerli domain ID
+//       providers: ["https://goerli.optimism.io"],
+//     },
+//   },
+// };
 
 export const ReplicaConfig = () => {
   // get chain id from useParams hook
@@ -76,15 +75,14 @@ export const ReplicaConfig = () => {
 
   return (
     <SingleColumnLayout title="Replicants">
-      <FormBuilder
+      {/* <FormBuilder
         form={APP_FORM.REPLICA_REGISTER}
         targetNetwork={TARGETS.NETWORK_ID}
         defaultValues={{
           chainID: chainID,
         }}
-      />
-      {/* This was the form to register and accept control on a forign chain. does not work because of preflight fail */}
-      {/* <FormBuilder
+      /> */}
+      <FormBuilder
         form={APP_FORM.BATCH_REPLICA_CLAIM}
         targetNetwork={TARGETS.NETWORK_ID}
         customFields={{ ...MolochFields, ...AppFieldLookup }}
@@ -93,7 +91,7 @@ export const ReplicaConfig = () => {
           chainID: chainID,
           domainID: domainID,
         }}
-      /> */}
+      />
     </SingleColumnLayout>
   );
 };
